Adding 125I Binding folder
The current idea is to separate 125I binding from 3H Binding, since the waste is handled differently and I believe we need to track more information for 125I ligands. This may not be the way we handle it in the future, but for right now this is what makes sense to me.