Starting to refactor Binding_Worksheet
Finally starting to refactor the binding_worksheet.py script to use the utils.py functions and clean up some messy names. Mainly restructured the beginning of the file but I hope to go through and condense more aspects of this. It currently should works so that alone is a victory

Also changed the rad waste name to be stored in a separate text file rather than as a variable defined in the script. I am starting to consider just making a config file for my script to read, that may be more useful instead of making a 3rd text file?